+++ tsup.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
     "src/v21.ts",
     "src/v22.ts",
     "src/v23.ts",
+    "src/v24.ts",
     "src/icons.ts",
   ],
   format: ["esm", "cjs"],